+++ talker/lib/src/extensions/log_level.dart
@@ -1,5 +1,4 @@
 import 'package:talker/talker.dart';
-import 'package:talker_error_handler/talker_error_handler.dart';
 
 extension ToLogLevel on ErrorLevel? {
   LogLevel get loglevel {