+++ talker/lib/src/models/talker_data/talker_log.dart
@@ -33,9 +33,7 @@ class TalkerLog implements TalkerDataInterface {
 
   @override
   String generateTextMessage() {
-    final mes = '$titleText$message$consoleAditional';
-    final underline = ConsoleFormater.getUnderLine(mes);
-    return mes + underline;
+    return '$titleText$message$displayAditional';
   }
 
   @override