+++ talker/lib/src/observers/talker_observer.dart
@@ -1,7 +1,12 @@
 import 'package:talker/src/models/talker_data/talker_data.dart';
 import 'package:talker_error_handler/talker_error_handler.dart';
 
-abstract class TalkerObserver {
-  Function(ErrorContainer err) get onError;
-  Function(TalkerDataInterface data) get onLog;
+class TalkerObserver {
+  const TalkerObserver({
+    required this.onError,
+    required this.onLog,
+  });
+
+  final Function(ErrorDetails err) onError;
+  final Function(TalkerDataInterface data) onLog;
 }