+++ talker_error_handler/lib/src/error_handler_interface.dart
@@ -1,26 +1,26 @@
 import 'package:talker_error_handler/talker_error_handler.dart';
 
 abstract class ErrorHandlerInterface {
-  Stream<ErrorContainer> get stream;
-  List<ErrorContainer> get history;
+  Stream<ErrorDetails> get stream;
+  List<ErrorDetails> get history;
 
-  ErrorContainer? handle(
-    String msg, [
-    Object? exception,
+  void handle(
+    Object exception, [
+    String? msg,
     StackTrace? stackTrace,
     ErrorLevel? errorLevel,
   ]);
 
-  ErrorContainer handleError(
-    String msg, [
-    Error? error,
+  void handleError(
+    Error error, [
+    String? msg,
     StackTrace? stackTrace,
     ErrorLevel? errorLevel,
   ]);
 
-  ErrorContainer handleException(
-    String msg, [
-    Exception? exception,
+  void handleException(
+    Exception exception, [
+    String? msg,
     StackTrace? stackTrace,
     ErrorLevel? errorLevel,
   ]);