+++ talker_error_handler/lib/src/models/error_details/error_details.dart
@@ -1,10 +1,10 @@
 //TODO: Container or wrapper
 import 'package:talker_error_handler/talker_error_handler.dart';
 
-export 'base_error_container.dart';
+export 'base_error_details.dart';
 
-abstract class ErrorContainer {
-  String get message;
+abstract class ErrorDetails {
+  String? get message;
   Exception? get exception;
   Error? get error;
   StackTrace? get stackTrace;