+++ talker_error_handler/lib/src/models/models.dart
@@ -1,3 +1,3 @@
-export 'error_container/error_container.dart';
+export 'error_details/error_details.dart';
 export 'error_handler_settings/error_handler_settings.dart';
 export 'error_level/error_level.dart';