+++ talker_logger/lib/src/talker_logger_base.dart
@@ -6,7 +6,7 @@ class TalkerLogger {
     TalkerLoggerFilter? filter,
     this.formater = const BaseLoggerFormater(),
   }) {
-    _filter = filter ?? BaseTalkerLoggerFilter(settings.logLevel);
+    _filter = filter ?? BaseTalkerLoggerFilter(settings.level);
   }
 
   final TalkerLoggerSettings settings;
@@ -16,8 +16,11 @@ class TalkerLogger {
   void log(String msg, {LogLevel logLevel = LogLevel.debug}) {
     var pen = settings.colors[logLevel];
     pen = pen ?? logLevel.consoleColor;
+
     if (_filter.shouldLog(msg, logLevel)) {
-      consolePrint(formater.fmt(pen.write(msg), logLevel));
+      consolePrint(
+        formater.fmt(msg, logLevel, pen),
+      );
     }
   }