+++ talker_logger/lib/src/utils/console_formater.dart
@@ -1,11 +1,13 @@
 abstract class ConsoleFormater {
   ConsoleFormater._();
 
-  static String getUnderLine(String mes) {
+  static String addUnderline(String mes) {
     final parts = mes.split('\n');
+
     final haveEndSep = parts.last.isEmpty;
     parts.sort((a, b) => a.length.compareTo(b.length));
     final lineLen = parts.last.length;
-    return haveEndSep ? '_' * lineLen : '\n${'_' * lineLen}';
+
+    return mes + (haveEndSep ? '_' * lineLen : '\n${'_' * lineLen}');
   }
 }