+++ talker_logger/lib/src/utils/utils.dart
@@ -1,2 +1,3 @@
 export 'console_formater.dart';
+export 'filter/filter.dart';
 export 'formater/formater.dart';