+++ talker_error_handler/CHANGELOG.md
@@ -1,3 +1,9 @@
-## 1.0.0
+## 0.2.0
+- Simplified package api
+- Add common handle method
+- Add onUnknownErrorType method
+
+
+## 0.1.0
 
 - Initial version.